Add reading & extracting values from counters data
Implement reading the counters data from the boiler EEPROM and
extracting the counter values from it.  This required a slight
adjustment to the Message::FormatNumber to allow increasing as well as
shrinking values when scaling them.

+++ Message.h
@@ -14,7 +14,7 @@ protected:
 
     std::string SpecialDataConv5(uint64_t data);
     std::string SpecialDataConv6(uint8_t data);
-    std::string FormatNumber(int value, uint scale, int min = INT_MIN, int max = INT_MAX, std::string minText = {}, std::string maxText = {});
+    std::string FormatNumber(int value, float scale, int min = INT_MIN, int max = INT_MAX, std::string minText = {}, std::string maxText = {});
     std::string NoYesCodeString(uint8_t value);
     std::string StatusCodeString(uint8_t value);
     std::string LockingCodeString(uint8_t value);
